Move NTP service tests to lib.rs
